Reformat Extensions, Add Koin Extension
