+++ 2005/developer-sungjae/ArchitectureStudy/app/src/main/java/com/project/architecturestudy/extensions/StringExt.kt
@@ -1,0 +1,9 @@
+package com.project.architecturestudy.extensions
+
+import org.apache.commons.text.StringEscapeUtils
+import org.jsoup.Jsoup
+
+fun String.parseHTMLTag(): String {
+    val strWithoutSign = StringEscapeUtils.unescapeHtml4(this)
+    return Jsoup.parse(strWithoutSign).text()
+}
