+++ 2005/developer-sungjae/ArchitectureStudy/app/src/main/java/com/project/architecturestudy/extensions/TextViewExt.kt
@@ -1,0 +1,9 @@
+package com.project.architecturestudy.extensions
+
+import android.widget.TextView
+import androidx.databinding.BindingAdapter
+
+@BindingAdapter("parseHtml")
+fun TextView.parseHtml(text: String?) {
+    this.text = text?.parseHTMLTag()
+}
